Polyfill VAO as function
